# Pre-compute values to speed up first search