--- a/src/enchcracker/SimpleRandom.java
+++ b/src/enchcracker/SimpleRandom.java
@@ -15,6 +15,16 @@ public class SimpleRandom {
         return (int)(seed >>> 17);
     }
 
+    // always used
+    public int next8() {
+        return (int)((8 * (long)next()) >> 31);
+    }
+
+    // specifically for 15 shelves
+    public int next8and16() {
+        return (int)((8 * (long)next()) >> 31) + (int)((16 * (long)next()) >> 31);
+    }
+
     public int nextInt(int bound) {
         int r = next();
         int m = bound - 1;